// types/address.d.ts
export interface Address {
    id: string
    user_id: string
    name_address: string
    city: string
    state: string
    postal_code: string
    country: string
    phone_number: string
    recipient_name: string
    address_line: string
    additional_address_line: string
    is_default: boolean
    created_at: string
    updated_at: string
}